Replace auto-trim feature in the image_converter
Change algorithm to search the most similar bit pattern of the top of
the track and the end of the track. Previously, the algorithm detects
IAMs and search the similar bit pattern around it because I thought the
bit pattens around IAMs have lot of changes (features) so it is easier
to find the similar pattern.